+++ pygeos/tests/test_io.py
@@ -7,7 +7,7 @@ import pytest
 
 import pygeos
 
-from .common import all_types, empty_point, point, point_z
+from .common import all_types, empty_point, empty_point_z, point, point_z
 
 # fmt: off
 POINT11_WKB = b"\x01\x01\x00\x00\x00" + struct.pack("<2d", 1.0, 1.0)
@@ -283,8 +283,21 @@ def test_to_wkt_geometrycollection_with_point_empty():
     assert pygeos.to_wkt(collection).endswith("(POINT EMPTY, POINT (2 3))")
 
 
+@pytest.mark.skipif(
+    pygeos.geos_version < (3, 9, 0),
+    reason="MULTIPOINT (EMPTY, 2 3) only works for GEOS >= 3.9",
+)
+def test_to_wkt_multipoint_with_point_empty():
+    geom = pygeos.multipoints([empty_point, point])
+    assert pygeos.to_wkt(geom) == "MULTIPOINT (EMPTY, 2 3)"
+
+
+@pytest.mark.skipif(
+    pygeos.geos_version >= (3, 9, 0),
+    reason="MULTIPOINT (EMPTY, 2 3) gives ValueError on GEOS < 3.9",
+)
 def test_to_wkt_multipoint_with_point_empty_errors():
-    # Test if segfault is prevented
+    # test if segfault is prevented
     geom = pygeos.multipoints([empty_point, point])
     with pytest.raises(ValueError):
         pygeos.to_wkt(geom)
@@ -302,6 +315,10 @@ def test_repr_max_length():
     assert representation.endswith("...>")
 
 
+@pytest.mark.skipif(
+    pygeos.geos_version >= (3, 9, 0),
+    reason="MULTIPOINT (EMPTY, 2 3) gives Exception on GEOS < 3.9",
+)
 def test_repr_multipoint_with_point_empty():
     # Test if segfault is prevented
     geom = pygeos.multipoints([point, empty_point])
@@ -374,88 +391,116 @@ def test_to_wkb_srid():
     assert np.frombuffer(result[5:9], "<u4").item() == 4326
 
 
-@pytest.mark.skipif(
-    pygeos.geos_version >= (3, 8, 0), reason="Pre GEOS 3.8.0 has 3D empty points"
-)
 @pytest.mark.parametrize(
-    "geom,dims,expected",
+    "geom,expected",
     [
-        (empty_point, 2, POINT_NAN_WKB),
-        (empty_point, 3, POINTZ_NAN_WKB),
-        (pygeos.multipoints([empty_point]), 2, MULTIPOINT_NAN_WKB),
-        (pygeos.multipoints([empty_point]), 3, MULTIPOINTZ_NAN_WKB),
-        (pygeos.geometrycollections([empty_point]), 2, GEOMETRYCOLLECTION_NAN_WKB),
-        (pygeos.geometrycollections([empty_point]), 3, GEOMETRYCOLLECTIONZ_NAN_WKB),
+        (empty_point, POINT_NAN_WKB),
+        (empty_point_z, POINT_NAN_WKB),
+        (pygeos.multipoints([empty_point]), MULTIPOINT_NAN_WKB),
+        (pygeos.multipoints([empty_point_z]), MULTIPOINT_NAN_WKB),
+        (pygeos.geometrycollections([empty_point]), GEOMETRYCOLLECTION_NAN_WKB),
+        (pygeos.geometrycollections([empty_point_z]), GEOMETRYCOLLECTION_NAN_WKB),
         (
             pygeos.geometrycollections([pygeos.multipoints([empty_point])]),
-            2,
             NESTED_COLLECTION_NAN_WKB,
         ),
         (
-            pygeos.geometrycollections([pygeos.multipoints([empty_point])]),
-            3,
+            pygeos.geometrycollections([pygeos.multipoints([empty_point_z])]),
+            NESTED_COLLECTION_NAN_WKB,
+        ),
+    ],
+)
+def test_to_wkb_point_empty_2d(geom, expected):
+    actual = pygeos.to_wkb(geom, output_dimension=2, byte_order=1)
+    # Split 'actual' into header and coordinates
+    coordinate_length = 16
+    header_length = len(expected) - coordinate_length
+    # Check the total length (this checks the correct dimensionality)
+    assert len(actual) == header_length + coordinate_length
+    # Check the header
+    assert actual[:header_length] == expected[:header_length]
+    # Check the coordinates (using numpy.isnan; there are many byte representations for NaN)
+    assert np.isnan(struct.unpack("<2d", actual[header_length:])).all()
+
+
+@pytest.mark.xfail(
+    pygeos.geos_version[:2] == (3, 8), reason="GEOS==3.8 never outputs 3D empty points"
+)
+@pytest.mark.parametrize(
+    "geom,expected",
+    [
+        (empty_point_z, POINTZ_NAN_WKB),
+        (pygeos.multipoints([empty_point_z]), MULTIPOINTZ_NAN_WKB),
+        (pygeos.geometrycollections([empty_point_z]), GEOMETRYCOLLECTIONZ_NAN_WKB),
+        (
+            pygeos.geometrycollections([pygeos.multipoints([empty_point_z])]),
             NESTED_COLLECTIONZ_NAN_WKB,
         ),
     ],
 )
-def test_to_wkb_point_empty_pre_geos38(geom, dims, expected):
-    actual = pygeos.to_wkb(geom, output_dimension=dims, byte_order=1)
-    # Use numpy.isnan; there are many byte representations for NaN
-    assert actual[: -dims * 8] == expected[: -dims * 8]
-    assert np.isnan(struct.unpack("<{}d".format(dims), actual[-dims * 8 :])).all()
+def test_to_wkb_point_empty_3d(geom, expected):
+    actual = pygeos.to_wkb(geom, output_dimension=3, byte_order=1)
+    # Split 'actual' into header and coordinates
+    coordinate_length = 24
+    header_length = len(expected) - coordinate_length
+    # Check the total length (this checks the correct dimensionality)
+    assert len(actual) == header_length + coordinate_length
+    # Check the header
+    assert actual[:header_length] == expected[:header_length]
+    # Check the coordinates (using numpy.isnan; there are many byte representations for NaN)
+    assert np.isnan(struct.unpack("<3d", actual[header_length:])).all()
 
 
-@pytest.mark.skipif(
-    pygeos.geos_version < (3, 8, 0), reason="Post GEOS 3.8.0 has 2D empty points"
+@pytest.mark.xfail(
+    pygeos.geos_version < (3, 8, 0),
+    reason="GEOS<3.8 always outputs 3D empty points if output_dimension=3",
 )
 @pytest.mark.parametrize(
-    "geom,dims,expected",
+    "geom,expected",
     [
-        (empty_point, 2, POINT_NAN_WKB),
-        (empty_point, 3, POINT_NAN_WKB),
-        (pygeos.multipoints([empty_point]), 2, MULTIPOINT_NAN_WKB),
-        (pygeos.multipoints([empty_point]), 3, MULTIPOINT_NAN_WKB),
-        (pygeos.geometrycollections([empty_point]), 2, GEOMETRYCOLLECTION_NAN_WKB),
-        (pygeos.geometrycollections([empty_point]), 3, GEOMETRYCOLLECTION_NAN_WKB),
+        (empty_point, POINT_NAN_WKB),
+        (pygeos.multipoints([empty_point]), MULTIPOINT_NAN_WKB),
+        (pygeos.geometrycollections([empty_point]), GEOMETRYCOLLECTION_NAN_WKB),
         (
             pygeos.geometrycollections([pygeos.multipoints([empty_point])]),
-            2,
-            NESTED_COLLECTION_NAN_WKB,
-        ),
-        (
-            pygeos.geometrycollections([pygeos.multipoints([empty_point])]),
-            3,
             NESTED_COLLECTION_NAN_WKB,
         ),
     ],
 )
-def test_to_wkb_point_empty_post_geos38(geom, dims, expected):
-    # Post GEOS 3.8: empty point is 2D
-    actual = pygeos.to_wkb(geom, output_dimension=dims, byte_order=1)
-    # Use numpy.isnan; there are many byte representations for NaN
-    assert actual[: -2 * 8] == expected[: -2 * 8]
-    assert np.isnan(struct.unpack("<2d", actual[-2 * 8 :])).all()
+def test_to_wkb_point_empty_2d_output_dim_3(geom, expected):
+    actual = pygeos.to_wkb(geom, output_dimension=3, byte_order=1)
+    # Split 'actual' into header and coordinates
+    coordinate_length = 16
+    header_length = len(expected) - coordinate_length
+    # Check the total length (this checks the correct dimensionality)
+    assert len(actual) == header_length + coordinate_length
+    # Check the header
+    assert actual[:header_length] == expected[:header_length]
+    # Check the coordinates (using numpy.isnan; there are many byte representations for NaN)
+    assert np.isnan(struct.unpack("<2d", actual[header_length:])).all()
 
 
 @pytest.mark.parametrize(
-    "wkb,expected_type",
+    "wkb,expected_type,expected_dim",
     [
-        (POINT_NAN_WKB, 0),
-        (POINTZ_NAN_WKB, 0),
-        (MULTIPOINT_NAN_WKB, 4),
-        (MULTIPOINTZ_NAN_WKB, 4),
-        (GEOMETRYCOLLECTION_NAN_WKB, 7),
-        (GEOMETRYCOLLECTIONZ_NAN_WKB, 7),
-        (NESTED_COLLECTION_NAN_WKB, 7),
-        (NESTED_COLLECTIONZ_NAN_WKB, 7),
+        (POINT_NAN_WKB, 0, 2),
+        (POINTZ_NAN_WKB, 0, 3),
+        (MULTIPOINT_NAN_WKB, 4, 2),
+        (MULTIPOINTZ_NAN_WKB, 4, 3),
+        (GEOMETRYCOLLECTION_NAN_WKB, 7, 2),
+        (GEOMETRYCOLLECTIONZ_NAN_WKB, 7, 3),
+        (NESTED_COLLECTION_NAN_WKB, 7, 2),
+        (NESTED_COLLECTIONZ_NAN_WKB, 7, 3),
     ],
 )
-def test_from_wkb_point_empty(wkb, expected_type):
+def test_from_wkb_point_empty(wkb, expected_type, expected_dim):
     geom = pygeos.from_wkb(wkb)
     # POINT (nan nan) transforms to an empty point
-    # Note that the dimensionality (2D/3D) is GEOS-version dependent
     assert pygeos.is_empty(geom)
     assert pygeos.get_type_id(geom) == expected_type
+    # The dimensionality (2D/3D) is only read correctly for GEOS >= 3.9.0
+    if pygeos.geos_version >= (3, 9, 0):
+        assert pygeos.get_coordinate_dimension(geom) == expected_dim
 
 
 def test_to_wkb_point_empty_srid():